Refactoring of tests:  encapsulating them into separate classes and replacing references to W3Schools with Google.com to improve load speeds
