+++ src/test/java/com/softwareonpurpose/uinavigator/UiElement4GetInstanceTests.java
@@ -1,0 +1,48 @@
+package com.softwareonpurpose.uinavigator;
+
+import org.testng.Assert;
+import org.testng.annotations.Test;
+
+@Test
+public class UiElement4GetInstanceTests {
+    @SuppressWarnings("rawtypes")
+    @Test
+    public void getInstance() {
+        Class expected = UiElement4.class;
+        String description = "'body' element";
+        String locatorType = UiLocatorType4.TAG;
+        String locatorValue = "body";
+        Class actual = UiElement4.getInstance(description, locatorType, locatorValue).getClass();
+        Assert.assertEquals(actual, expected);
+    }
+
+    @SuppressWarnings("rawtypes")
+    @Test(enabled = false)
+    public void getInstance_ordinal() {
+        Class expected = UiElement4.class;
+        String description = "'paragraph' element";
+        String locatorType = UiLocatorType4.TAG;
+        String locatorValue = "p";
+        int ordinal = 2;
+        Class actual = UiElement4.getInstance(description, locatorType, locatorValue, ordinal).getClass();
+        Assert.assertEquals(actual, expected);
+    }
+
+    @SuppressWarnings("rawtypes")
+    @Test(enabled = false)
+    public void getInstance_parent() {
+        Class expected = UiElement4.class;
+        UiElement4 parent = UiElement4.getInstance("'Unordered List' element", UiLocatorType4.TAG, "ul");
+        Class actual = UiElement4.getInstance("'List Item' element", UiLocatorType4.TAG, "li", parent).getClass();
+        Assert.assertEquals(actual, expected);
+    }
+
+    @SuppressWarnings("rawtypes")
+    @Test(enabled = false)
+    public void getInstance_ordinal_parent() {
+        Class expected = UiElement4.class;
+        UiElement4 parent = UiElement4.getInstance("'Ordered List' element", UiLocatorType4.TAG, "ol");
+        Class actual = UiElement4.getInstance("'List Item' element", UiLocatorType4.TAG, "li", 2, parent).getClass();
+        Assert.assertEquals(actual, expected);
+    }
+}
