+++ src/test/resources/link.html
@@ -5,7 +5,7 @@
 <h2>HTML Links</h2>
 <p>HTML links are defined with the a tag:</p>
 
-<a href="https://www.w3schools.com">This is a link</a>
+<a href="https://www.google.com">This is a link</a>
 
 </body>
 </html>
